@@ -10,16 +10,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^11.0.3",
+    "@angular/animations": "^11.2.0",
     "@angular/cdk": "^11.0.3",
-    "@angular/common": "~11.0.3",
-    "@angular/compiler": "~11.0.3",
-    "@angular/core": "~11.0.3",
-    "@angular/forms": "~11.0.3",
+    "@angular/common": "~11.2.0",
+    "@angular/compiler": "~11.2.0",
+    "@angular/core": "~11.2.0",
+    "@angular/forms": "~11.2.0",
     "@angular/material": "^11.0.2",
-    "@angular/platform-browser": "~11.0.3",
-    "@angular/platform-browser-dynamic": "~11.0.3",
-    "@angular/router": "~11.0.3",
+    "@angular/platform-browser": "~11.2.0",
+    "@angular/platform-browser-dynamic": "~11.2.0",
+    "@angular/router": "~11.2.0",
     "primeicons": "^4.1.0",
     "primeng": "^11.1.0-rc.1",
     "rxjs": "~6.6.3",
@@ -27,17 +27,17 @@
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.1100.3",
-    "@angular/cli": "^11.1.2",
-    "@angular/compiler-cli": "~11.0.3",
-    "@angular/language-service": "~11.0.3",
+    "@angular-devkit/build-angular": "^0.1102.0",
+    "@angular/cli": "^11.2.0",
+    "@angular/compiler-cli": "~11.2.0",
+    "@angular/language-service": "~11.2.0",
     "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
     "codelyzer": "^6.0.0",
     "jasmine-core": "~3.6.0",
     "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~5.0.0",
+    "karma": "~6.1.1",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage-istanbul-reporter": "~3.0.2",
     "karma-jasmine": "~4.0.0",
