@@ -2,7 +2,7 @@ import { Component, ChangeDetectionStrategy, OnDestroy } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
-import { AgentsDao } from 'src/app/services/agents-dao.service';
+import { AgentsDao } from 'src/app/shared/services/agents-dao.service';
 import { Agent } from 'src/app/shared/models';
 
 @Component({
@@ -16,14 +16,15 @@ export class AgentPageComponent implements OnDestroy {
 
   agent: Agent;
 
-  constructor(private route: ActivatedRoute,
-    private router: Router,
-    private titleService: Title,
-    private agentsDao: AgentsDao) {
+  constructor(private readonly route: ActivatedRoute,
+    private readonly router: Router,
+    private readonly titleService: Title,
+    private readonly agentsDao: AgentsDao) {
     this.subscription = this.route.params.subscribe((params: Params) => {
       if (params.agent) {
-        this.agent = this.agentsDao.getAgentByFullName(
-          params.agent.split('-').join(' '));
+        const fullName = params.agent.includes('ilek-nowak') ?
+          'Magdalena Ilek-Nowak' : params.agent.split('-').join(' ');
+        this.agent = this.agentsDao.getAgentByFullName(fullName);
         if (this.agent) {
           this.titleService.setTitle(this.agent.fullName);
         } else {
