@@ -1,5 +1,5 @@
 import { Component, HostListener, OnDestroy, OnInit } from '@angular/core';
-import { WindowSizeDetector } from 'src/app/services/window-size-detector.service';
+import { WindowSizeDetector } from 'src/app/shared/services/window-size-detector.service';
 import { trigger, style, animate, transition } from '@angular/animations';
 import { PrimeNGConfig } from 'primeng/api';
 import { ABOUT_US_LINKS, ALL_LINKS } from './header/menu-links';
@@ -42,8 +42,8 @@ export class AppComponent implements OnInit, OnDestroy {
   aboutUsLinks = ABOUT_US_LINKS;
 
   constructor(readonly windowSizeDetector: WindowSizeDetector,
-    private titleService: Title, private router: Router,
-    private primengConfig: PrimeNGConfig) {
+    private readonly titleService: Title, private readonly router: Router,
+    private readonly primengConfig: PrimeNGConfig) {
       this. subscription = router.events.subscribe(event => {
         if(event instanceof NavigationEnd) {
           if (!this.isUrlTitleComputedInComponent(event.url)) {
