@@ -1,14 +1,14 @@
-import { Component, ChangeDetectionStrategy, ChangeDetectorRef, OnDestroy } from '@angular/core';
+import { Component, ChangeDetectionStrategy, OnInit, ChangeDetectorRef } from '@angular/core';
+import { OnDestroy } from '@angular/core';
 import { Title } from '@angular/platform-browser';
 import { ActivatedRoute, Params, Router } from '@angular/router';
 import { Subscription } from 'rxjs';
-import { OffersDao } from '../services/offers-dao.service';
-import { Offer, GalleryPhoto } from '../shared/models';
-import { AgentsDao } from 'src/app/services/agents-dao.service';
+import { OffersDao } from '../shared/services/offers-dao.service';
+import { Offer, OfferField } from '../shared/models';
+import { AgentsDao } from 'src/app/shared/services/agents-dao.service';
 import { Agent } from 'src/app/shared/models';
-import { SnackbarService } from '../services/snackbar.service';
-import { WindowSizeDetector } from '../services/window-size-detector.service'
-import { DESKTOP_LARGE, DESKTOP_SMALL, MOBILE, TABLET } from '../services/window-size-detector.service';
+import { SnackbarService } from '../shared/services/snackbar.service';
+import { WindowSizeDetector } from '../shared/services/window-size-detector.service'
 
 @Component({
   selector: 'perfect-offer',
@@ -16,43 +16,32 @@ import { DESKTOP_LARGE, DESKTOP_SMALL, MOBILE, TABLET } from '../services/window
   styleUrls: ['./offer.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
-export class OfferComponent implements OnDestroy {
+export class OfferComponent implements OnInit, OnDestroy {
   private subscription: Subscription;
   offer: Offer;
-  images: GalleryPhoto[] = [];
-  responsiveOptions: any[] = [
-    {
-        breakpoint: DESKTOP_LARGE + 'px',
-        numVisible: 5
-    },
-    {
-        breakpoint: DESKTOP_SMALL + 'px',
-        numVisible: 4
-    },
-    {
-        breakpoint: TABLET + 'px',
-        numVisible: 3
-    },
-    {
-        breakpoint: MOBILE + 'px',
-        numVisible: 1
-    }
-];
-  constructor(private route: ActivatedRoute,
-    private changeDetector: ChangeDetectorRef,
-    private router: Router,
-    private titleService: Title,
-    private offersDao: OffersDao,
-    readonly agentsDao: AgentsDao,
+  definedOfferFields: OfferField<any>[] = [];
+
+  constructor(readonly agentsDao: AgentsDao,
     readonly windowSizeDetector: WindowSizeDetector,
-    private snackbarService: SnackbarService) {
-    this.subscription = this.route.params.subscribe((params: Params) => {
+    private readonly route: ActivatedRoute,
+    private readonly changeDetector: ChangeDetectorRef,
+    private readonly router: Router,
+    private readonly titleService: Title,
+    private readonly offersDao: OffersDao,
+    private readonly snackbarService: SnackbarService) {
+      this.subscription = this.windowSizeDetector.windowSizeChanged$.subscribe(() => {
+        this.changeDetector.detectChanges();
+      });
+    }
+
+  ngOnInit() {
+    this.subscription.add(this.route.params.subscribe((params: Params) => {
         if (params.symbol) {
           this.loadOffer(params.symbol);
         } else  {
           this.handleNonexistentOffer();
         }    
-      });
+      }));
   }
 
   private loadOffer(symbol: string) {
@@ -62,7 +51,7 @@ export class OfferComponent implements OnDestroy {
       this.offer = this.offersDao.getOfferBySymbol(symbol);
       if (this.offer) {
         this.titleService.setTitle(this.offer.title);
-        this.images = this.computeImages(this.offer.photos);
+        this.definedOfferFields = this.computeDefinedOfferFields();
         this.changeDetector.detectChanges();
       } else {
         this.handleNonexistentOffer(symbol);
@@ -83,12 +72,34 @@ export class OfferComponent implements OnDestroy {
     }
   }
   
-  handleNonexistentOffer(symbol = '') {
+  private handleNonexistentOffer(symbol = '') {
     setTimeout(() => {
       this.snackbarService.open(`Oferta ${symbol} nie istnieje.`);
     }, 500)
     this.router.navigate(['/oferty']);
   }
+
+  private computeDefinedOfferFields(): OfferField<any>[] {
+    return Object.values(this.offer).filter(value => this.isDefinedOfferField(value)); 
+  }
+
+  private isDefinedOfferField(value: any): boolean {
+    return this.isOfferField(value) && this.isDefined(value.value);
+  }
+
+  // Because of https://stackoverflow.com/a/46703380/11212568 instanceof
+  // does not work on interface.
+  private isOfferField(value: any): boolean {
+    return value.hasOwnProperty('displayName') && value.hasOwnProperty('value');
+  }
+
+  private isDefined(value: any): boolean {
+    if (typeof value === 'number') {
+    return value > -1;
+    }
+    return !!value;  
+  }
+
   navigateToAgentPage(agent: Agent) {
     this.router.navigate(['/ludzie/' + this.computeAgentLink(agent.fullName)]);
   }
@@ -96,17 +107,7 @@ export class OfferComponent implements OnDestroy {
   computeAgentLink(agentFullName: string): string {
     return agentFullName.toLowerCase().split(' ').join('-');
   }
-  
-  computeImages(photos: string[]): GalleryPhoto[] {
-    let computedPhotos: GalleryPhoto[] = [];
-    for (let photo of photos) {
-      computedPhotos.push({ 
-        previewImageSrc: '/offers/' + photo,
-        thumbnailImageSrc: '/offers/' + photo,
-      })
-    }
-    return computedPhotos;
-  }
+
   ngOnDestroy() {
     this.subscription.unsubscribe();
   }
