@@ -1,8 +1,8 @@
 export * from "./agent";
-export * from "./estate";
+export * from "./estate-type";
 export * from "./filters";
 export * from "./offer";
+export * from "./offer-field";
 export * from "./room";
 export * from "./sorting";
-export * from "./transaction";
 export * from "./gallery-photo";
