@@ -1,8 +1,10 @@
+import { OfferField } from "./offer-field";
 import { Room } from "./room";
 
 export interface Offer {
     id: number;
     estateType: string;
+    estateSubtypes: string[];
     agentId: number;
     title: string;
     description: string;
@@ -13,9 +15,9 @@ export interface Offer {
     number: number;
     symbol: string;
     status: string;
-    legalStatus: string;
-    standard: number;
-    isAlarmSet: boolean;
+    legalStatus: OfferField<string>;
+    standard: OfferField<number>;
+    isAlarmSet: OfferField<boolean>;
     isExclusive: boolean;
     isForRent: boolean
     photos: string[];
@@ -28,15 +30,15 @@ export interface Offer {
     virtualVisitUrl: string;
 
     price: number;
-    pricePerSquareMeter: number;
-    pricePerUsableSquareMeter: number;
-    rentPrice: number;
-    grossRentPricePerSquareMeter: number;
-    netRentPricePerSquareMeter: number;
-    deposit: number;
-    depositType: string;
+    pricePerSquareMeter: OfferField<number>;
+    pricePerUsableSquareMeter: OfferField<number>;
+    rentPrice: OfferField<number>;
+    grossRentPricePerSquareMeter: OfferField<number>;
+    netRentPricePerSquareMeter: OfferField<number>;
+    deposit: OfferField<number>;
+    depositType: OfferField<string>;
     additionalFees: string[];
-    garagePrice: number;
+    garagePrice: OfferField<number>;
 
     country: string;
     voivodeship: string;
@@ -50,81 +52,79 @@ export interface Offer {
     mapLongtitude: string;
     mapLatitude: string;
     publicTransport: string[];
-    neighbourhood: string;
-    accessRoad: string;
-    distance: number;
+    neighbourhood: OfferField<string>;
+    accessRoad: OfferField<string>;
 
-    totalArea: number;
-    landArea: number;
-    usableArea: number;
-    terraceArea: number;
-    officeArea: number;
-    officeBuildingArea: number;
-    floorArea: number;
+    totalArea: OfferField<number>;
+    landArea: OfferField<number>;
+    usableArea: OfferField<number>;
+    terraceArea: OfferField<number>;
+    officeArea: OfferField<number>;
+    officeBuildingArea: OfferField<number>;
+    floorArea: OfferField<number>;
 
-    garage: string;
-    buildingName: string;
-    gasConnectionDetails: string; 
-    numberOfRooms: number;
-    numberOfBedrooms: number;
-    numberOfTerraces: number;
-    numberOfFloors: number;
-    floor: number;
-    floorHeight: number;
-    isBasementAvailable: boolean;
-    sewageConnectionDetails: string; 
-    burden: string,
-    heatingType: string;
-    windowsType: string;
+    garage: OfferField<string>;
+    buildingName: OfferField<string>;
+    gasConnectionDetails: OfferField<string>; 
+    numberOfRooms: OfferField<number>;
+    numberOfBedrooms: OfferField<number>;
+    numberOfTerraces: OfferField<number>;
+    numberOfFloors: OfferField<number>;
+    floor: OfferField<number>;
+    floorHeight: OfferField<number>;
+    isBasementAvailable: OfferField<boolean>;
+    sewageConnectionDetails: OfferField<string>; 
+    burden: OfferField<string>,
+    heatingType: OfferField<string>;
+    windowsType: OfferField<string>;
     isMarketPrimary: boolean;
-    acquisitionType: string;
+    acquisitionType: OfferField<string>;
     rooms: Room[];
-    officeBuildingCategory: string;
-    electricityConnectionDetails: string;
-    buildingType: string;
-    flatType: string;
-    yearBuilt: number;
-    isTerraceAvailable: boolean;
-    flatSetUp: string[];
-    waterConnectionDetails: string; 
-    roomsHeight: number;
-    isParkingAvailable: boolean;
-    isFenceAvailable: boolean;
-    isElevatorAvailable: boolean;
-    isAccessible: boolean;
-    minimumRentingPeriodInMonths: number;
-    predestination: string[];
-    isReceptionAvailable: boolean;
-    isComputerNetworkAvailable: boolean;
-    securityType: string;
-    entrance: string;
-    noiseLevel: string;
-    constructionTechnology: string;
+    officeBuildingCategory: OfferField<string>;
+    electricityConnectionDetails: OfferField<string>;
+    buildingType: OfferField<string>;
+    flatType: OfferField<string>;
+    yearBuilt: OfferField<number>;
+    isTerraceAvailable: OfferField<boolean>;
+    flatSetUp: OfferField<string>;
+    waterConnectionDetails: OfferField<string>; 
+    roomsHeight: OfferField<number>;
+    isParkingAvailable: OfferField<boolean>;
+    isFenceAvailable: OfferField<boolean>;
+    isElevatorAvailable: OfferField<boolean>;
+    isAccessible: OfferField<boolean>;
+    minimumRentingPeriodInMonths: OfferField<number>;
+    isReceptionAvailable: OfferField<boolean>;
+    isComputerNetworkAvailable: OfferField<boolean>;
+    securityType: OfferField<string>;
+    entrance: OfferField<string>;
+    noiseLevel: OfferField<string>;
+    constructionTechnology: OfferField<string>;
     // Fields related to halls:
-    isFloorDustFree: boolean;
-    isLoadingRampAvailable: boolean;
-    warehouseArea: number;
-    warehouseYardArea: number;
-    isYardHardened: boolean;
-    isManeuveringAreaAvailable: boolean;
-    isTruckParkingAvailable: boolean;
-    isGantryAvailable: boolean;
-    isGoodsLiftAvailable: boolean;
+    isFloorDustFree: OfferField<boolean>;
+    isLoadingRampAvailable: OfferField<boolean>;
+    warehouseArea: OfferField<number>;
+    warehouseYardArea: OfferField<number>;
+    isYardHardened: OfferField<boolean>;
+    isManeuveringAreaAvailable: OfferField<boolean>;
+    isTruckParkingAvailable: OfferField<boolean>;
+    isGantryAvailable: OfferField<boolean>;
+    isGoodsLiftAvailable: OfferField<boolean>;
     // Fields related to premises:
-    premiseState: string;
-    premiseType: string;
-    premiseBackRoomArea: number;
-    premiseSalesRoomArea: number;
+    premiseState: OfferField<string>;
+    premiseType: OfferField<string>;
+    premiseBackRoomArea: OfferField<number>;
+    premiseSalesRoomArea: OfferField<number>;
     // Fields related to land lots:
-    mpzp: string;
-    mpzpInfo: string;
-    landLotUse: string;
-    landLotDetailedSize: string;
-    landLotConstructionConditions: string;
-    landLotForm: string;
-    landLotFrontLineWidth: number;
-    landLotShape: string;
-    landLotLegalStatus: string;
-    landLotFenceType: string;
-    landLotBuildings: string;
+    mpzp: OfferField<string>;
+    mpzpInfo: OfferField<string>;
+    landLotUse: OfferField<string>;
+    landLotDetailedSize: OfferField<string>;
+    landLotConstructionConditions: OfferField<string>;
+    landLotForm: OfferField<string>;
+    landLotFrontLineWidth: OfferField<number>;
+    landLotShape: OfferField<string>;
+    landLotLegalStatus: OfferField<string>;
+    landLotFenceType: OfferField<string>;
+    landLotBuildings: OfferField<string>;
 }
